UI enhancements to the Control.java and MenuBar.java classes.
Also, this may or may not be a good idea, but I felt it was fitting: I made the exit_button about twice as wide as the other buttons. Not only does it make it easier to hit, but it fills up the space that was empty before (see for yourself). 

Oh, silly me, I used absolute values for relative pointers in networkengage.java (so I corrected them to the proper R.drawable.bla format). 

Some style enhancements for the networkengage.java file too, including a dim radial gradient that uses the shape xml system (so no bitmaps!!). 

One version up in Manifest too.

-Georgiy