Fixed the issues with Custom Element Maker causing a Force Close (Issue 14 and Issue 19).
What I did was make it so that rather than having a save file for the JNI, which only had the collisions saved and a save file for the Java with everything saved, I only used the Java save file, made DemoActivity call CustomMaker.loadCustom() to initialize the JNI with the custom data at the beginning each time, and made sure no other calls were made to saving the JNI version. The only file that is now used is the Java one, which is now called customproperties.txt, though that can be changed too. I also added some checking for if the file doesn't exist, because it seemed this caused errors too, but it may have been redundant, I'm just too dead to try to figure it out right now.

Minor other thing was that I moved the intro note into strings.xml and made it only appear on the first run. We may want to change what it says at some point, and this makes it easy to find.